+++ src/components/atom/pagination/Pagination.test.js
@@ -6,12 +6,14 @@ import Pagination from './Pagination';
 
 describe('Pagination', () => {
   it('should match with the snapshot', () => {
-    const wrapper = shallow(<Pagination
-            activePage={1}
-            totalItemsCount={450}
-            pageRangeDisplayed={5}
-            onChange={() => {}}
-          />);
+    const wrapper = shallow(
+      <Pagination
+        activePage={1}
+        totalItemsCount={450}
+        pageRangeDisplayed={5}
+        onChange={() => {}}
+      />
+    );
     expect(toJson(wrapper)).toMatchSnapshot();
   });
 });