+++ src/components/molecule/form-group/FormGroup.js
@@ -4,33 +4,25 @@ import PropTypes from 'prop-types';
 
 import css from './FormGroup.module.scss';
 
-const FormGroup = ({
-  success,
-  error,
-  children
-}) => {
+const FormGroup = ({ success, error, children }) => {
   const classes = classNames({
     [css['c-form-group']]: true,
     [css['c-form-group--error']]: error,
     [css['c-form-group--success']]: success
   });
 
-  return (
-    <div className={classes}>
-      {children}
-    </div>
-  );
+  return <div className={classes}>{children}</div>;
 };
 
 FormGroup.propTypes = {
   error: PropTypes.bool,
   success: PropTypes.bool,
-  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number])
+  children: PropTypes.instanceOf(Array)
 };
 FormGroup.defaultProps = {
   error: false,
   success: false,
-  value: null
+  children: null
 };
 
 export default FormGroup;