+++ src/docs/stories/11-Heading.stories.js
@@ -16,4 +16,6 @@ export const H3 = () => <Heading level={3} title="Title" />;
 export const H4 = () => <Heading level={4} title="Title" />;
 export const H5 = () => <Heading level={5} title="Title" />;
 export const H6 = () => <Heading level={6} title="Title" />;
-export const withOtherProps = () => <Heading level={6} title="Title" style={someStyle} />;
+export const withOtherProps = () => (
+  <Heading level={6} title="Title" style={someStyle} />
+);