+++ src/docs/stories/9-Pagination.stories.js
@@ -1,34 +1,19 @@
-import React, { useState } from 'react';
-import Pagination from '../../components/molecule/pagination/Pagination';
+import React from 'react';
 import { action } from '@storybook/addon-actions';
+import Pagination from '../../components/atom/pagination/Pagination';
 
 export default {
   title: 'Pagination',
   component: Pagination
 };
 
-export const Default = () => {
-  function Parent({ children, ...props }) {
-    const [state, setState] = useState();
-    return <div>{children(state, setState)}</div>;
-  }
-
-
-  return (
-    <Parent>
-      {(state, setState) => {
-        state = state || {page: 1};
-        return (
-          <Pagination
-            activePage={state.page}
-            itemsCountPerPage={10}
-            totalItemsCount={450}
-            pageRangeDisplayed={5}
-            onClick={action('clicked')}
-            onChange={page => setState({ page })}
-          />
-        )
-      }}
-    </Parent>
-  )
-};
+export const Default = () => (
+  <Pagination
+    activePage={1}
+    itemsCountPerPage={10}
+    totalItemsCount={450}
+    pageRangeDisplayed={5}
+    onClick={action('clicked')}
+    onChange={action('changePage')}
+  />
+);