+++ src/index.js
@@ -4,13 +4,18 @@ import { Grid, Col, Row } from 'react-flexbox-grid';
 // UI
 // atoms
 import Button from './components/atom/button/Button';
+import DatePicker from './components/atom/datepicker/DatePicker';
+import Heading from './components/atom/heading/Heading';
 import Input from './components/atom/input/Input';
 import Label from './components/atom/label/Label';
+import Pagination from './components/atom/pagination/Pagination';
+import Select from './components/atom/select/Select';
 
 // molecules
 import FormActions from './components/molecule/form-actions/FormActions';
 import FormGroup from './components/molecule/form-group/FormGroup';
 import FormRow from './components/molecule/form-row/FormRow';
+
 
 // organisms
 import Form from './components/organism/form/Form';
@@ -18,12 +23,16 @@ import Form from './components/organism/form/Form';
 export {
   Button,
   Col,
+  DatePicker,
   Form,
   FormActions,
   FormGroup,
   FormRow,
   Grid,
+  Heading,
   Input,
   Label,
+  Pagination,
+  Select,
   Row
 };